+++ canon/Cargo.toml
@@ -13,11 +13,11 @@ license = "MPL-2.0"
 blake2b_simd = { version = "0.3", default-features = false }
 arrayvec = { version = "0.5", default-features = false }
 arbitrary = { version = "0.3", features = ["derive"], optional = true }
+cfg-if = "1.0.0"
 
 [dev-dependencies]
 canonical_host = { path = "../canon_host", version = "0.4.0" }
 canonical_derive = { path = "../canon_derive", version = "0.4.0" }
-canonical_collections = { path = "../canon_collections", version = "0.4.0" }
 canonical_fuzz = { path = "../canon_fuzz", version = "0.4.0" }
 
 [features]
