+++ canon/src/lib.rs
@@ -10,20 +10,15 @@
 #![deny(missing_docs)]
 
 mod canon;
+mod repr;
+
+pub use repr::Repr;
 
 #[cfg(not(feature = "host"))]
 mod bridge;
-#[cfg(not(feature = "host"))]
-mod repr_hosted;
+
 #[cfg(not(feature = "host"))]
 pub use bridge::BridgeStore;
-#[cfg(not(feature = "host"))]
-pub use repr_hosted::Repr;
-
-#[cfg(feature = "host")]
-mod repr_host;
-#[cfg(feature = "host")]
-pub use repr_host::{Repr, ValMut};
 
 mod cow;
 mod dry_sink;
@@ -32,7 +27,7 @@ mod implementations;
 mod store;
 
 pub use canon::{Canon, InvalidEncoding};
-pub use cow::Cow;
+pub use cow::{Cow, CowMut};
 pub use dry_sink::DrySink;
 pub use id::Id32;
 pub use store::{ByteSink, ByteSource, IdBuilder, Ident, Sink, Source, Store};
