--- canon_collections/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "canonical_collections"
-version = "0.4.0"
-authors = ["Kristoffer Ström <kristoffer@dusk.network>"]
-edition = "2018"
-readme = "README.md"
-repository = "https://github.com/dusk-network/canonical/canon_collections"
-keywords = ["canon", "serialisation", "no_std", "ffi", "database"]
-description = "Data structures for use with Canonical"
-license = "MPL-2.0"
-
-[dependencies]
-canonical_derive = { path = "../canon_derive", version = "0.4.0" }
-canonical = { path = "../canon", version = "0.4.0" }
-
-[dev-dependencies]
-canonical_host = { path = "../canon_host", version = "0.4.0" }
-
-[features]
-host = ["canonical/host"]
