--- canon_collections/README.md
@@ -1,3 +1,0 @@
-# Canonical collections
-
-Collections on canon, used for tests atm.
