+++ canon_host/Cargo.toml
@@ -13,8 +13,8 @@ license = "MPL-2.0"
 lazy_static = "1.4"
 parking_lot = "0.11"
 wasmi = "0.6.0"
-canonical = { path = "../canon", version = "0.4.0" }
-canonical_derive = { path = "../canon_derive", version = "0.4.0" }
+canonical = "0.4.0"
+canonical_derive = "0.4.0"
 
 [dev-dependencies]
 counter = { path = "examples/counter", features = ["host"], version = "0.1.0" }
